requirements-deploy.txt
fastapi>=0.111.0
uvicorn>=0.29.0
motor>=3.4.0
pymongo[srv]>=4.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
openai>=1.30.0
faiss-cpu>=1.8.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
numpy>=1.26.0
requests>=2.31.0
python-dateutil>=2.9.0
